# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EVS API",
    "description": "USACE Enterprise Visibility Suite API.\n\nEvery endpoint that replaces an Oracle APEX page carries an `x-apex-page` extension so the\ntraceability matrix (legacy/traceability.csv) can be regenerated from the contract.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/programs": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Programs (APEX Initiatives IR)",
        "description": "Interactive Report parameters: repeatable `filter=column:op:value` (op: eq, ne, gt, gte, lt, lte, like, in), `q` full text, `sort=-col,col2`, `limit`, `offset`, `format=csv`. Columns: program_code, name, business_line, division, funded_amount, obligated_amount, expended_amount, variance_pct, project_count, schedule_health.",
        "operationId": "list_programs_api_v1_programs_get",
        "parameters": [
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "business_line",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Line"
            }
          },
          {
            "name": "schedule_health",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schedule Health"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated columns, `-` prefix for descending",
              "title": "Sort"
            },
            "description": "Comma separated columns, `-` prefix for descending"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text search across text columns (IR search bar)",
              "title": "Q"
            },
            "description": "Free text search across text columns (IR search bar)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "`csv` streams the filtered rows",
              "default": "json",
              "title": "Format"
            },
            "description": "`csv` streams the filtered rows"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramList"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "20",
        "x-apex-authorization": "Authenticated User"
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Projects (APEX Projects IR)",
        "description": "Interactive Report parameters: repeatable `filter=column:op:value` (op: eq, ne, gt, gte, lt, lte, like, in), `q` full text, `sort=-col,col2`, `limit`, `offset`, `format=csv`. Columns: p2_project_no, name, program_code, district, division, business_line, phase, pdt_lead, baseline_finish, current_finish, pct_complete, funded_amount, obligated_amount, expended_amount, schedule_health.",
        "operationId": "list_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "program_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Code"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "business_line",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Line"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phase"
            }
          },
          {
            "name": "schedule_health",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schedule Health"
            }
          },
          {
            "name": "min_pct_complete",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Pct Complete"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated columns, `-` prefix for descending",
              "title": "Sort"
            },
            "description": "Comma separated columns, `-` prefix for descending"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text search across text columns (IR search bar)",
              "title": "Q"
            },
            "description": "Free text search across text columns (IR search bar)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "`csv` streams the filtered rows",
              "default": "json",
              "title": "Format"
            },
            "description": "`csv` streams the filtered rows"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "30",
        "x-apex-authorization": "Authenticated User"
      }
    },
    "/api/v1/projects/{p2_project_no}": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Project detail (APEX Project Details)",
        "operationId": "get_project_api_v1_projects__p2_project_no__get",
        "parameters": [
          {
            "name": "p2_project_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "P2 Project No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "3",
        "x-apex-process": "log",
        "x-apex-authorization": "Authenticated User"
      }
    },
...
    "/api/v1/schedule/milestones": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "Milestones with slip (status=slipped for the late list)",
        "operationId": "milestones_api_v1_schedule_milestones_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "complete",
                    "scheduled",
                    "slipped"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "program_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Code"
            }
          },
          {
            "name": "p2_project_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "P2 Project No"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma separated columns, `-` prefix for descending",
              "title": "Sort"
            },
            "description": "Comma separated columns, `-` prefix for descending"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text search across text columns (IR search bar)",
              "title": "Q"
            },
            "description": "Free text search across text columns (IR search bar)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv)$",
              "description": "`csv` streams the filtered rows",
              "default": "json",
              "title": "Format"
            },
            "description": "`csv` streams the filtered rows"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneList"
                }
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-apex-page": "4",
        "x-apex-authorization": "Authenticated User"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AsOf": {
        "properties": {
          "source_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source As Of",
            "description": "Newest timestamp inside the source payload"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At",
            "description": "When the EVS ingestion worker stored it"
          },
          "freshness": {
            "type": "string",
            "enum": [
              "fresh",
              "aging",
              "stale",
              "simulated"
            ],
            "title": "Freshness",
            "default": "fresh"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "live | fixtures | simulated | synthetic",
            "default": "fixtures"
          }
        },
        "type": "object",
        "title": "AsOf",
        "description": "Every dataset carries both the source timestamp and when EVS fetched it."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Milestone": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "baseline_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Date"
          },
          "current_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Date"
          },
          "actual_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "scheduled",
              "slipped"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "baseline_date",
          "current_date",
          "actual_date",
          "status"
        ],
        "title": "Milestone"
      },
      "MilestoneList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectMilestone"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "as_of"
        ],
        "title": "MilestoneList"
      },
...
      "Page": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "title": "Page"
      },
      "Program": {
        "properties": {
          "program_code": {
            "type": "string",
            "title": "Program Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "business_line": {
            "type": "string",
            "title": "Business Line"
          },
          "division": {
            "type": "string",
            "title": "Division"
          },
          "funded_amount": {
            "type": "number",
            "title": "Funded Amount"
          },
          "obligated_amount": {
            "type": "number",
            "title": "Obligated Amount"
          },
          "expended_amount": {
            "type": "number",
            "title": "Expended Amount"
          },
          "variance_pct": {
            "type": "number",
            "title": "Variance Pct"
          },
          "project_count": {
            "type": "integer",
            "title": "Project Count"
          },
          "schedule_health": {
            "type": "string",
            "enum": [
              "on_track",
              "at_risk",
              "late"
            ],
            "title": "Schedule Health"
          }
        },
        "type": "object",
        "required": [
          "program_code",
          "name",
          "business_line",
          "division",
          "funded_amount",
          "obligated_amount",
          "expended_amount",
          "variance_pct",
          "project_count",
          "schedule_health"
        ],
        "title": "Program"
      },
      "ProgramList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Program"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "as_of"
        ],
        "title": "ProgramList"
      },
...
      "Project": {
        "properties": {
          "p2_project_no": {
            "type": "string",
            "title": "P2 Project No"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "program_code": {
            "type": "string",
            "title": "Program Code"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "division": {
            "type": "string",
            "title": "Division"
          },
          "business_line": {
            "type": "string",
            "title": "Business Line"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "pdt_lead": {
            "type": "string",
            "title": "Pdt Lead"
          },
          "baseline_finish": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Finish"
          },
          "current_finish": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Finish"
          },
          "pct_complete": {
            "type": "number",
            "title": "Pct Complete"
          },
          "funded_amount": {
            "type": "number",
            "title": "Funded Amount"
          },
          "obligated_amount": {
            "type": "number",
            "title": "Obligated Amount"
          },
          "expended_amount": {
            "type": "number",
            "title": "Expended Amount"
          },
          "schedule_health": {
            "type": "string",
            "enum": [
              "on_track",
              "at_risk",
              "late"
            ],
            "title": "Schedule Health"
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/Milestone"
            },
            "type": "array",
            "title": "Milestones",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "p2_project_no",
          "name",
          "program_code",
          "district",
          "division",
          "business_line",
          "phase",
          "pdt_lead",
          "baseline_finish",
          "current_finish",
          "pct_complete",
          "funded_amount",
          "obligated_amount",
          "expended_amount",
          "schedule_health"
        ],
        "title": "Project"
      },
...
      "ProjectList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "as_of": {
            "$ref": "#/components/schemas/AsOf"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "as_of"
        ],
        "title": "ProjectList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}